feature(toggle theme) enable button functionality to toggle theme
