+++ client/app.js
@@ -1,21 +1,17 @@
 import React from 'react'
-import CssBaseline from '@material-ui/core/CssBaseline'
-import {createMuiTheme, ThemeProvider} from '@material-ui/core/styles'
-import theme from './theme/theme'
 import Routes from './routes'
-import CustomThemeProvider from './components/CustomMUI/CustomThemeProvider'
-const dark = createMuiTheme({
-  palette: {
-    type: 'dark'
-  }
-})
+import CustomThemeProvider from './theme/components/CustomThemeProvider'
+import CssBaseline from '@material-ui/core/CssBaseline'
+import {ThemeProvider} from '@material-ui/core/styles'
+import theme from './theme/baseTheme'
+
 export default function App() {
   return (
-    <div>
-      <ThemeProvider theme={theme}>
+    <ThemeProvider theme={theme}>
+      <CustomThemeProvider>
         <CssBaseline />
         <Routes />
-      </ThemeProvider>
-    </div>
+      </CustomThemeProvider>
+    </ThemeProvider>
   )
 }
