+++ client/components/index.js
@@ -1,7 +1,5 @@
-/**
- * `components/index.js` exists simply as a 'central export' for our components.
- * This way, we can import all of our components from the same place, rather than
- * having to figure out which file they belong to!
+/*
+`components/index.js` exists as a 'central export' for our components.
  */
 
 export {default as Navbar} from './Navbar'
