+++ client/theme/normalTheme.js
@@ -1,28 +1,28 @@
 import {createMuiTheme} from '@material-ui/core/styles'
-import {red} from '@material-ui/core/colors'
 import {fonts, backupFonts} from './fonts'
-// Dark theme
+//default 'normal' theme
 const theme = createMuiTheme({
   palette: {
-    type: 'dark',
     primary: {
-      main: 'rgba(187, 161, 193, 1)',
-      light: 'rgb(81, 91, 95)',
-      dark: 'rgb(26, 35, 39)',
-      contrastText: '#ffffff'
+      main: '#000000'
     },
     secondary: {
-      light: 'rgba(252, 248, 244, 1)',
-      main: 'rgba(187, 161, 193, 1)',
-      dark: 'rgba(98, 18, 68, 1)',
-      contrastText: 'rgba(255, 255, 255, 1)'
-    },
-    titleBar: {
-      main: '#555555',
-      contrastText: '#ffffff'
+      main: '#FEF6D3'
     },
     error: {
-      main: red.A400
+      main: '#C10230',
+      dark: '#C10230',
+      contrastText: '#ffffff'
+    },
+    text: {
+      primary: '#000000',
+      secondary: '#000000',
+      disabled: '#A69B95',
+      hint: '#FABA91'
+    },
+    background: {
+      paper: '#ffffff',
+      default: '#ffffff'
     }
   },
   typography: {
