+++ client/theme/wildcardTheme.js
@@ -1,28 +1,18 @@
 import {createMuiTheme} from '@material-ui/core/styles'
-import {red} from '@material-ui/core/colors'
 import {fonts, backupFonts} from './fonts'
-// Dark theme
+// 'wildcard' theme
 const theme = createMuiTheme({
   palette: {
-    type: 'dark',
     primary: {
-      main: 'rgba(187, 161, 193, 1)',
-      light: 'rgb(81, 91, 95)',
-      dark: 'rgb(26, 35, 39)',
-      contrastText: '#ffffff'
+      main: '#77FF33'
     },
     secondary: {
-      light: 'rgba(252, 248, 244, 1)',
-      main: 'rgba(187, 161, 193, 1)',
-      dark: 'rgba(98, 18, 68, 1)',
-      contrastText: 'rgba(255, 255, 255, 1)'
-    },
-    titleBar: {
-      main: '#555555',
-      contrastText: '#ffffff'
+      main: '#FEF6D3'
     },
     error: {
-      main: red.A400
+      main: '#C10230',
+      dark: '#C10230',
+      contrastText: '#ffffff'
     }
   },
   typography: {
